rename 'render' option to 'renderCombination'
