+++ src/renderCombination.js
